backend/requirements.txt
fastapi>=0.123.0
uvicorn[standard]>=0.38.0
python-multipart>=0.0.20
pydantic>=2.12.0
pypdf>=6.4.0
python-docx>=1.2.0
